+++ components/Header/Navigator/ProfileButton.module.scss
@@ -28,7 +28,7 @@
       width: 26px;
       word-break: keep-all;
       color: $grayscale-gray4;
-      margin: auto;
+      margin-left: -1px;
     }
   }
 }
