+++ components/PostDetailPage/PostDetail/Comments/Comment.module.scss
@@ -26,7 +26,7 @@
     }
 
     .userNickname {
-      height: 100%;
+      height: 20px;
       position: absolute;
       left: 40px;
       top: 8px;
