+++ components/PostDetailPage/PostDetail/Comments/CommentForm.module.scss
@@ -1,0 +1,49 @@
+@import "../../../../assets/colors/colors";
+@import "../../../../assets/fonts/typography_medium";
+@import "../../../../assets/fonts/typography_regular";
+
+.container {
+  width: 551px;
+  height: 151px;
+  margin: 12px 0 0 288px;
+  background: $grayscale-white;
+  position: relative;
+
+  textarea {
+    width: 519px;
+    height: 80px;
+    position: absolute;
+    top: 16px;
+    left: 16px;
+    padding: 10px 8px;
+    color: #666666;
+    overflow: auto;
+    resize: none;
+    border: none;
+
+    &::placeholder {
+      color: $grayscale-gray5;
+    }
+  }
+
+  .addPhoto {
+    position: absolute;
+    top: 112px;
+    left: 16px;
+
+    svg {
+      path {
+        fill: $grayscale-gray3;
+      }
+    }
+  }
+
+  .submit {
+    height: 26px;
+    width: 30px;
+    position: absolute;
+    top: 108px;
+    left: 497px;
+    color: $purple-core;
+  }
+}
