+++ components/PostDetailPage/PostDetail/PostDetail.tsx
@@ -1,5 +1,6 @@
 import PostWriteButton from "../../PostsPage/Posts/WriteAPost/PostWriteButton/PostWriteButton";
 import BoardName from "./BoardName";
+import CommentForm from "./Comments/CommentForm";
 import CommentsList from "./Comments/CommentsList";
 import PostMain from "./PostMain";
 
@@ -10,6 +11,7 @@ export default function PostDetail() {
       <PostWriteButton postDetailPage />
       <PostMain />
       <CommentsList />
+      <CommentForm />
     </>
   );
 }
